+++ e2e/test-environment-esm/__tests__/testUsingMjsEnv.test.js
@@ -1,0 +1,13 @@
+/**
+ * Copyright (c) Meta Platforms, Inc. and affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ * @jest-environment <rootDir>/env-4.mjs
+ */
+'use strict';
+
+test('should pass', () => {
+  expect(globalThis.someVar).toBe(42);
+});