+++ e2e/test-environment-esm/__tests__/testUsingMtsEnv.test.js
@@ -3,13 +3,11 @@
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
+ *
+ * @jest-environment <rootDir>/env-1.mts
  */
+'use strict';
 
-import {TestEnvironment} from 'jest-environment-node';
-
-export default class Env extends TestEnvironment {
-  constructor(...args) {
-    super(...args);
-    this.global.someVar = 42;
-  }
-}
+test('should pass', () => {
+  expect(globalThis.someVar).toBe(42);
+});