+++ e2e/test-environment-esm/__tests__/testUsingTsEnv.test.js
@@ -3,9 +3,11 @@
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
+ *
+ * @jest-environment <rootDir>/env-2.ts
  */
 'use strict';
 
-test('dummy', () => {
+test('should pass', () => {
   expect(globalThis.someVar).toBe(42);
 });