+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - `[jest-reporters]` Fix issue where console output not displayed for GHA reporter even with `silent: false` option ([#15864](https://github.com/jestjs/jest/pull/15864))
 - `[jest-runtime]` Fix issue where user cannot utilize dynamic import despite specifying `--experimental-vm-modules` Node option ([#15842](https://github.com/jestjs/jest/pull/15842))
 - `[jest-test-sequencer]` Fix issue where failed tests due to compilation errors not getting re-executed even with `--onlyFailures` CLI option ([#15851](https://github.com/jestjs/jest/pull/15851))
+- `[jest-runner, jest-transformer]` support custom test environment written in ESM with `TypeScript` extensions ([#15885](https://github.com/jestjs/jest/pull/15885)), fixes downstream [issue](https://github.com/kulshekhar/ts-jest/issues/4195)
 
 ### Chore & Maintenance